acheived a clean build with -Wall, but will not use -Werror for things may change with compiler updates

+++ src/Sequencer.cpp
@@ -184,7 +184,7 @@ void Sequencer::ChangeActivePattern(int new_one){
 }
 
 void Sequencer::ClearPattern(int p){
-    if (p >= patterns.size()) return;
+    if (p >= (int)patterns.size()) return;
     patterns[p].Clear();
 
 }
